benchmark/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 3D processing
trimesh>=4.3.0
pyrender>=0.1.45

# CLIP
git+https://github.com/openai/CLIP.git
image-reward>=1.5

# Image processing
Pillow>=10.0.0
imageio>=2.31.0
scikit-image>=0.22.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0

# Data handling
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Progress bars
tqdm>=4.65.0

# Utilities
pyyaml>=6.0